Fixed order of wrappers in order for Auth and  API context to have access to router
