+++ src/frontend/src/routes.tsx
@@ -11,6 +11,8 @@ import { ProtectedRoute } from "./components/authGuard";
 import { ProtectedLoginRoute } from "./components/authLoginGuard";
 import { AuthSettingsGuard } from "./components/authSettingsGuard";
 import { StoreGuard } from "./components/storeGuard";
+import ContextWrapper from "./contexts";
+import { AppInitPage } from "./pages/AppInitPage";
 import { AppWrapperPage } from "./pages/AppWrapperPage";
 import { DashboardWrapperPage } from "./pages/DashboardWrapperPage";
 import FlowPage from "./pages/FlowPage";
@@ -35,129 +37,146 @@ const PlaygroundPage = lazy(() => import("./pages/Playground"));
 const SignUp = lazy(() => import("./pages/SignUpPage"));
 const router = createBrowserRouter(
   createRoutesFromElements([
-    <Route path="/" element={<AppWrapperPage />}>
-      <Route
-        path=""
-        element={
-          <ProtectedRoute>
-            <Outlet />
-          </ProtectedRoute>
-        }
-      >
-        <Route path="" element={<DashboardWrapperPage />}>
-          <Route path="" element={<HomePage />}>
-            <Route index element={<Navigate replace to={"all"} />} />
-            <Route
-              path="flows/"
-              element={<MyCollectionComponent key="flows" type="flow" />}
-            >
+    <Route
+      path="/"
+      element={
+        <ContextWrapper>
+          <Outlet />
+        </ContextWrapper>
+      }
+    >
+      <Route path="" element={<AppInitPage />}>
+        <Route path="" element={<AppWrapperPage />}>
+          <Route
+            path=""
+            element={
+              <ProtectedRoute>
+                <Outlet />
+              </ProtectedRoute>
+            }
+          >
+            <Route path="" element={<DashboardWrapperPage />}>
+              <Route path="" element={<HomePage />}>
+                <Route index element={<Navigate replace to={"all"} />} />
+                <Route
+                  path="flows/"
+                  element={<MyCollectionComponent key="flows" type="flow" />}
+                >
+                  <Route
+                    path="folder/:folderId"
+                    element={<MyCollectionComponent key="flows" type="flow" />}
+                  />
+                </Route>
+                <Route
+                  path="components/"
+                  element={
+                    <MyCollectionComponent key="components" type="component" />
+                  }
+                >
+                  <Route
+                    path="folder/:folderId"
+                    element={
+                      <MyCollectionComponent
+                        key="components"
+                        type="component"
+                      />
+                    }
+                  />
+                </Route>
+                <Route
+                  path="all/"
+                  element={<MyCollectionComponent key="all" type="all" />}
+                >
+                  <Route
+                    path="folder/:folderId"
+                    element={<MyCollectionComponent key="all" type="all" />}
+                  />
+                </Route>
+              </Route>
+              <Route path="/settings" element={<SettingsPage />}>
+                <Route index element={<Navigate replace to={"general"} />} />
+                <Route
+                  path="global-variables"
+                  element={<GlobalVariablesPage />}
+                />
+                <Route path="api-keys" element={<ApiKeysPage />} />
+                <Route
+                  path="general/:scrollId?"
+                  element={
+                    <AuthSettingsGuard>
+                      <GeneralPage />
+                    </AuthSettingsGuard>
+                  }
+                />
+                <Route path="shortcuts" element={<ShortcutsPage />} />
+                <Route path="messages" element={<MessagesPage />} />
+              </Route>
               <Route
-                path="folder/:folderId"
-                element={<MyCollectionComponent key="flows" type="flow" />}
-              />
-            </Route>
-            <Route
-              path="components/"
-              element={
-                <MyCollectionComponent key="components" type="component" />
-              }
-            >
-              <Route
-                path="folder/:folderId"
+                path="/store"
                 element={
-                  <MyCollectionComponent key="components" type="component" />
+                  <StoreGuard>
+                    <StorePage />
+                  </StoreGuard>
+                }
+              />
+              <Route
+                path="/store/:id/"
+                element={
+                  <StoreGuard>
+                    <StorePage />
+                  </StoreGuard>
+                }
+              />
+              <Route path="/account">
+                <Route path="delete" element={<DeleteAccountPage />}></Route>
+              </Route>
+              <Route
+                path="/admin"
+                element={
+                  <ProtectedAdminRoute>
+                    <AdminPage />
+                  </ProtectedAdminRoute>
                 }
               />
             </Route>
-            <Route
-              path="all/"
-              element={<MyCollectionComponent key="all" type="all" />}
-            >
-              <Route
-                path="folder/:folderId"
-                element={<MyCollectionComponent key="all" type="all" />}
-              />
+            <Route path="/flow/:id/">
+              <Route path="" element={<DashboardWrapperPage />}>
+                <Route path="folder/:folderId/" element={<FlowPage />} />
+                <Route path="" element={<FlowPage />} />
+              </Route>
+              <Route path="view" element={<ViewPage />} />
+            </Route>
+            <Route path="/playground/:id/">
+              <Route path="" element={<PlaygroundPage />} />
             </Route>
           </Route>
-          <Route path="/settings" element={<SettingsPage />}>
-            <Route index element={<Navigate replace to={"general"} />} />
-            <Route path="global-variables" element={<GlobalVariablesPage />} />
-            <Route path="api-keys" element={<ApiKeysPage />} />
-            <Route
-              path="general/:scrollId?"
-              element={
-                <AuthSettingsGuard>
-                  <GeneralPage />
-                </AuthSettingsGuard>
-              }
-            />
-            <Route path="shortcuts" element={<ShortcutsPage />} />
-            <Route path="messages" element={<MessagesPage />} />
-          </Route>
           <Route
-            path="/store"
+            path="/login"
             element={
-              <StoreGuard>
-                <StorePage />
-              </StoreGuard>
+              <ProtectedLoginRoute>
+                <LoginPage />
+              </ProtectedLoginRoute>
             }
           />
           <Route
-            path="/store/:id/"
+            path="/signup"
             element={
-              <StoreGuard>
-                <StorePage />
-              </StoreGuard>
+              <ProtectedLoginRoute>
+                <SignUp />
+              </ProtectedLoginRoute>
             }
           />
-          <Route path="/account">
-            <Route path="delete" element={<DeleteAccountPage />}></Route>
-          </Route>
           <Route
-            path="/admin"
+            path="/login/admin"
             element={
-              <ProtectedAdminRoute>
-                <AdminPage />
-              </ProtectedAdminRoute>
+              <ProtectedLoginRoute>
+                <LoginAdminPage />
+              </ProtectedLoginRoute>
             }
           />
-        </Route>
-        <Route path="/flow/:id/">
-          <Route path="" element={<DashboardWrapperPage />}>
-            <Route path="folder/:folderId/" element={<FlowPage />} />
-            <Route path="" element={<FlowPage />} />
-          </Route>
-          <Route path="view" element={<ViewPage />} />
-        </Route>
-        <Route path="/playground/:id/">
-          <Route path="" element={<PlaygroundPage />} />
+          <Route path="*" element={<Navigate replace to="/" />} />
         </Route>
       </Route>
-      <Route
-        path="/login"
-        element={
-          <ProtectedLoginRoute>
-            <LoginPage />
-          </ProtectedLoginRoute>
-        }
-      />
-      <Route
-        path="/signup"
-        element={
-          <ProtectedLoginRoute>
-            <SignUp />
-          </ProtectedLoginRoute>
-        }
-      />
-      <Route
-        path="/login/admin"
-        element={
-          <ProtectedLoginRoute>
-            <LoginAdminPage />
-          </ProtectedLoginRoute>
-        }
-      />
-      <Route path="*" element={<Navigate replace to="/" />} />
     </Route>,
   ]),
 );
